VICTORY Works!
Added P to the maze - Peach’s starting point.

Removed the B point and the Polygon.

created variable to store current location (apart from OG location) and
draw now returns that variable to the variable.

Calling Mario and Peach with the OG location when creating them.

Changed the victory condition to when their locations meet.